+++ swift_browser_ui/sharing/server.py
@@ -12,6 +12,7 @@ import uvloop
 import swift_browser_ui.common.common_handlers
 import swift_browser_ui.common.common_middleware
 import swift_browser_ui.common.common_util
+import swift_browser_ui.common.db
 from swift_browser_ui.sharing.api import (
     access_details_handler,
     delete_share_handler,
@@ -27,25 +28,12 @@ from swift_browser_ui.sharing.api import (
     share_container_handler,
     shared_details_handler,
 )
-from swift_browser_ui.sharing.db import DBConn
 
 logging.basicConfig(level=logging.DEBUG)
 
 # temporarily ignore typecheck from mypy until
 # this issue is fixed https://github.com/MagicStack/uvloop/issues/575
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())  # type: ignore
-
-
-async def resume_on_start(app: aiohttp.web.Application) -> None:
-    """Resume old instance from start."""
-    app["db_conn"] = DBConn()
-    await app["db_conn"].open()
-
-
-async def save_on_shutdown(app: aiohttp.web.Application) -> None:
-    """Flush the database on shutdown."""
-    if app["db_conn"] is not None:
-        await app["db_conn"].close()
 
 
 async def init_server() -> aiohttp.web.Application:
@@ -59,6 +47,9 @@ async def init_server() -> aiohttp.web.Application:
             swift_browser_ui.common.common_middleware.error_handler,  # type: ignore
         ]
     )
+
+    # Cache db class
+    app["db_class"] = swift_browser_ui.common.db.SharingDBConn
 
     async def on_prepare(
         _: aiohttp.web.Request, response: aiohttp.web.StreamResponse
@@ -109,9 +100,9 @@ async def init_server() -> aiohttp.web.Application:
         ]
     )
 
-    app.on_startup.append(resume_on_start)
+    app.on_startup.append(swift_browser_ui.common.db.db_graceful_start)
     app.on_startup.append(swift_browser_ui.common.common_util.read_in_keys)
-    app.on_shutdown.append(save_on_shutdown)
+    app.on_shutdown.append(swift_browser_ui.common.db.db_graceful_close)
 
     return app